refactor: create the directory AddEditForm
