+++ src/features/foods/components/index.ts
@@ -1,8 +1,8 @@
 export { AddButton } from './AddButton';
-export { AddDrawerContent } from './AddDrawerContent';
+export { AddDrawerContent } from './AddEditForm/AddDrawerContent';
+export { CategorySelect } from './AddEditForm/CategorySelect';
 export { BadgeList } from './BadgeList';
 export { CategoryBadge } from './CategoryBadge';
-export { CategorySelect } from './CategorySelect';
 export { FilterBadge } from './FilterBadge';
 export { FilterButton } from './FilterButton';
 export { FoodList } from './FoodList';
